refactor(filid): split ast modules + mcp tools (test-metrics/doc-compress/fractal-navigate/lca-resolve)
Batch B: 9 files split into single-fn files with facade preservation.

- ast/ast-grep-shared (5 fn) → 4 files in utils/
- ast/cyclomatic-complexity (4 top + 16 inline) → 3 files in utils/
- ast/lcom4 (3 top + 7 inline) → 1 helper in utils/
- ast/dependency-extractor (3 top) → 2 files in utils/
- ast/tree-diff (3 top) → 2 files in utils/
- mcp/tools/test-metrics (4 fn) → 3 files
- mcp/tools/doc-compress (4 fn) → 3 files
- mcp/tools/fractal-navigate (4 fn) → 3 files
- mcp/tools/lca-resolve (3 fn) → 2 files
