+++ packages/filid/src/ast/ast-grep-shared/utils/get-files-for-language.ts
@@ -1,0 +1,52 @@
+import { readdirSync, statSync } from 'node:fs';
+import { extname, join, resolve } from 'node:path';
+
+import { AST_MAX_FILES, AST_SKIP_DIRS, EXT_TO_LANG } from '../../../constants/ast-languages.js';
+
+export function getFilesForLanguage(
+  dirPath: string,
+  language: string,
+  maxFiles = AST_MAX_FILES,
+): string[] {
+  const files: string[] = [];
+  const extensions = Object.entries(EXT_TO_LANG)
+    .filter(([_, lang]) => lang === language)
+    .map(([ext]) => ext);
+
+  function walk(dir: string) {
+    if (files.length >= maxFiles) return;
+
+    try {
+      const entries = readdirSync(dir, { withFileTypes: true });
+      for (const entry of entries) {
+        if (files.length >= maxFiles) return;
+
+        const fullPath = join(dir, entry.name);
+
+        // Skip common non-source directories
+        if (entry.isDirectory()) {
+          if (!AST_SKIP_DIRS.includes(entry.name)) {
+            walk(fullPath);
+          }
+        } else if (entry.isFile()) {
+          const ext = extname(entry.name).toLowerCase();
+          if (extensions.includes(ext)) {
+            files.push(fullPath);
+          }
+        }
+      }
+    } catch {
+      // Ignore permission errors
+    }
+  }
+
+  const resolvedPath = resolve(dirPath);
+  const stat = statSync(resolvedPath);
+
+  if (stat.isFile()) {
+    return [resolvedPath];
+  }
+
+  walk(resolvedPath);
+  return files;
+}
