+++ packages/filid/src/ast/ast-grep-shared/utils/to-lang-enum.ts
@@ -1,0 +1,16 @@
+import type { NapiLang, SgModule } from '../../../types/index.js';
+
+import { getMappedLang } from './getMappedLang.js';
+
+/**
+ * Convert lowercase language string to ast-grep Lang enum value
+ */
+export function toLangEnum(sg: SgModule, language: string): NapiLang {
+  // Lang enum only contains built-in languages (Html, JavaScript, Tsx, Css, TypeScript).
+  // All others (Python, Go, Rust, etc.) are CustomLang strings passed directly.
+  const lang = getMappedLang(sg, language);
+  if (lang === undefined) {
+    throw new Error(`Unsupported language: ${language}`);
+  }
+  return lang;
+}
