+++ packages/filid/src/ast/cyclomatic-complexity/cyclomatic-complexity.ts
@@ -9,54 +9,10 @@ import type { SgNode } from '@ast-grep/napi';
 
 import type { CyclomaticComplexityResult } from '../../types/metrics.js';
 
-import { parseSource, walk } from '../parser/parser.js';
-import { DECISION_KINDS } from '../../constants/decision-points.js';
-
-function computeCC(bodyNode: SgNode): number {
-  let cc = 1;
-  walk(bodyNode, (node: SgNode) => {
-    const kind = node.kind();
-
-    if (DECISION_KINDS.has(kind)) {
-      cc++;
-      return;
-    }
-
-    // switch_case: only non-default cases (those with a test expression)
-    if (kind === 'switch_case') {
-      // A default case in tree-sitter has kind 'switch_default'
-      // switch_case always has a condition
-      cc++;
-      return;
-    }
-
-    // switch_default is separate, don't count it
-    // (tree-sitter separates switch_case and switch_default)
-
-    // Logical && and || operators
-    if (kind === 'binary_expression') {
-      const children = node.children();
-      for (const child of children) {
-        const text = child.text();
-        if (text === '&&' || text === '||') {
-          cc++;
-          break;
-        }
-      }
-    }
-  });
-  return cc;
-}
-
-function getNodeName(node: SgNode): string | null {
-  const children = node.children();
-  // function_declaration: name is an identifier child
-  const nameNode = children.find(
-    (c: SgNode) =>
-      c.kind() === 'identifier' || c.kind() === 'property_identifier',
-  );
-  return nameNode ? nameNode.text() : null;
-}
+import { parseSource } from '../parser/parser.js';
+import { computeCC } from './utils/compute-cc.js';
+import { getNodeName } from './utils/get-node-name.js';
+import { processClassMethods } from './utils/process-class-methods.js';
 
 export async function calculateCC(
   source: string,
@@ -176,28 +132,4 @@ export async function calculateCC(
   for (const cc of perFunction.values()) fileTotal += cc;
 
   return { value: fileTotal, perFunction, fileTotal };
-}
-
-function processClassMethods(
-  classNode: SgNode,
-  perFunction: Map<string, number>,
-): void {
-  const classBody = classNode
-    .children()
-    .find((c: SgNode) => c.kind() === 'class_body');
-  if (!classBody) return;
-
-  for (const member of classBody.children()) {
-    if (member.kind() === 'method_definition') {
-      const nameNode = member
-        .children()
-        .find((c: SgNode) => c.kind() === 'property_identifier');
-      const body = member
-        .children()
-        .find((c: SgNode) => c.kind() === 'statement_block');
-      if (nameNode && body) {
-        perFunction.set(nameNode.text(), computeCC(body));
-      }
-    }
-  }
 }
