+++ packages/filid/src/ast/dependency-extractor/utils/strip-quotes.ts
@@ -1,0 +1,3 @@
+export function stripQuotes(s: string): string {
+  return s.replace(/^['"]|['"]$/g, '');
+}
