+++ packages/filid/src/ast/tree-diff/utils/get-name-from-node.ts
@@ -1,0 +1,6 @@
+import type { SgNode } from '@ast-grep/napi';
+
+export function getNameFromNode(node: SgNode, nameKind: string): string | null {
+  const nameNode = node.children().find((c: SgNode) => c.kind() === nameKind);
+  return nameNode ? nameNode.text() : null;
+}
