+++ packages/filid/src/mcp/tools/doc-compress/utils/handle-lossy.ts
@@ -1,0 +1,12 @@
+import { summarizeLossy } from '../../../../compress/lossy-summarizer/lossy-summarizer.js';
+import type { DocCompressInput, DocCompressOutput } from '../doc-compress.js';
+
+export function handleLossy(input: DocCompressInput): DocCompressOutput {
+  const entries = input.toolCallEntries ?? [];
+  const result = summarizeLossy(entries);
+
+  return {
+    summary: result.summary,
+    meta: result.meta,
+  };
+}
