+++ packages/filid/src/mcp/tools/fractal-navigate/fractal-navigate.ts
@@ -1,13 +1,9 @@
-import {
-  type NodeEntry,
-  buildFractalTree,
-  findNode,
-} from '../../../core/tree/fractal-tree/fractal-tree.js';
-import {
-  type ClassifyInput,
-  classifyNode,
-} from '../../../core/tree/organ-classifier/organ-classifier.js';
+import type { NodeEntry } from '../../../core/tree/fractal-tree/fractal-tree.js';
 import type { FractalTree } from '../../../types/fractal.js';
+
+import { handleClassify } from './utils/handle-classify.js';
+import { handleSiblingList } from './utils/handle-sibling-list.js';
+import { handleTree } from './utils/handle-tree.js';
 
 /** Input for fractal-navigate tool */
 export interface FractalNavigateInput {
@@ -52,76 +48,4 @@ export function handleFractalNavigate(
     default:
       return { error: `Unknown action: ${input.action}` };
   }
-}
-
-function handleClassify(input: FractalNavigateInput): FractalNavigateOutput {
-  // First check if the entry already has a known type in the provided entries
-  const entry = input.entries.find((e) => e.path === input.path);
-  if (entry && entry.type !== ('directory' as string)) {
-    return { classification: entry.type };
-  }
-
-  const dirName =
-    input.path
-      .split('/')
-      .filter((s) => s.length > 0)
-      .pop() ?? '';
-  const hasIntentMd = entry?.hasIntentMd ?? false;
-  const hasDetailMd = entry?.hasDetailMd ?? false;
-  const hasIndex = entry?.hasIndex ?? false;
-
-  // entries에서 실제 계산 (hardcode 제거)
-  const immediateChildren = input.entries.filter(
-    (e) =>
-      e.path !== input.path &&
-      e.path.startsWith(input.path + '/') &&
-      e.path.replace(input.path + '/', '').indexOf('/') === -1,
-  );
-  const hasFractalChildren = immediateChildren.some(
-    (c) => c.type === 'fractal' || c.type === 'pure-function',
-  );
-  const isLeafDirectory = immediateChildren.length === 0;
-
-  const classifyInput: ClassifyInput = {
-    dirName,
-    hasIntentMd,
-    hasDetailMd,
-    hasFractalChildren,
-    isLeafDirectory,
-    hasIndex,
-  };
-  const result = classifyNode(classifyInput);
-  return { classification: result };
-}
-
-function handleSiblingList(input: FractalNavigateInput): FractalNavigateOutput {
-  const tree = buildFractalTree(input.entries);
-
-  // Find parent path
-  const lastSlash = input.path.lastIndexOf('/');
-  const parentPath = lastSlash > 0 ? input.path.slice(0, lastSlash) : '';
-  const parentNode = findNode(tree, parentPath);
-
-  if (!parentNode) {
-    return { siblings: [] };
-  }
-
-  // children and organs are string[] (paths), resolve them to node info
-  const allChildPaths = [...parentNode.children, ...parentNode.organs];
-  const siblings = allChildPaths
-    .filter((p) => p !== input.path)
-    .map((p) => {
-      const node = tree.nodes.get(p);
-      return node
-        ? { name: node.name, path: node.path, type: node.type }
-        : { name: p.split('/').pop() ?? '', path: p, type: 'directory' };
-    });
-
-  return { siblings };
-}
-
-function handleTree(input: FractalNavigateInput): FractalNavigateOutput {
-  const tree = buildFractalTree(input.entries);
-  tree.nodesList = Array.from(tree.nodes.values());
-  return { tree };
 }
