+++ packages/filid/src/mcp/tools/fractal-navigate/utils/handle-tree.ts
@@ -1,0 +1,8 @@
+import { buildFractalTree } from '../../../../core/tree/fractal-tree/fractal-tree.js';
+import type { FractalNavigateInput, FractalNavigateOutput } from '../fractal-navigate.js';
+
+export function handleTree(input: FractalNavigateInput): FractalNavigateOutput {
+  const tree = buildFractalTree(input.entries);
+  tree.nodesList = Array.from(tree.nodes.values());
+  return { tree };
+}
