+++ packages/filid/src/mcp/tools/lca-resolve/utils/distance-between.ts
@@ -1,0 +1,7 @@
+import { pathDepth } from './path-depth.js';
+
+export function distanceBetween(from: string, to: string): number {
+  const fromDepth = pathDepth(from);
+  const toDepth = pathDepth(to);
+  return Math.abs(fromDepth - toDepth);
+}
