+++ packages/filid/src/mcp/tools/lca-resolve/utils/path-depth.ts
@@ -1,0 +1,3 @@
+export function pathDepth(p: string): number {
+  return p.split('/').filter((s) => s.length > 0).length;
+}
